refactor: group direct image ingestion entrypoint
