+++ apps/backend/src/entrypoints/directImageIngestion/lambda.ts
@@ -1,7 +1,7 @@
 import { handle } from "hono/aws-lambda";
-import { createDirectImageIngestionApp } from "../server/mediaRequests/directImageIngestionApp";
-import { getAllowedBrowserOrigins } from "../server/browserCors";
-import { createDirectImageIngestionLambdaHandler } from "./directImageIngestionLambdaHandler";
+import { createDirectImageIngestionApp } from "../../server/mediaRequests/directImageIngestionApp";
+import { getAllowedBrowserOrigins } from "../../server/browserCors";
+import { createDirectImageIngestionLambdaHandler } from "./handler";
 
 const handleRequest = handle(createDirectImageIngestionApp());
