+++ infra/aws/lib/gateways/api-gateway.ts
@@ -668,7 +668,8 @@ function createDirectImageIngestionFunction(
       "backend",
       "src",
       "entrypoints",
-      "lambda-direct-image-ingestion.ts",
+      "directImageIngestion",
+      "lambda.ts",
     ),
     handler: "handler",
     runtime: lambda.Runtime.NODEJS_24_X,
